+++ src/main/java/com/mercuriusxeno/goo/item/GooFormat.java
@@ -7,17 +7,17 @@ package com.mercuriusxeno.goo.item;
 public final class GooFormat {
 
     /** Volume threshold: sub-blob microblobs (<1,000). */
-    private static final int THRESHOLD_MICROBLOB = 1_000;
+    private static final long THRESHOLD_MICROBLOB = 1_000;
     /** Volume threshold: blob tier (<1 million mB). */
-    private static final int THRESHOLD_BLOB = 1_000_000;
+    private static final long THRESHOLD_BLOB = 1_000_000;
     /** Volume threshold: kiloblob tier (<1 billion mB). */
-    private static final int THRESHOLD_KILO = 1_000_000_000;
+    private static final long THRESHOLD_KILO = 1_000_000_000;
     /** Divisor for blob tier. */
-    private static final int DIVISOR_BLOB = 1_000;
+    private static final long DIVISOR_BLOB = 1_000;
     /** Divisor for kiloblob tier. */
-    private static final int DIVISOR_KILO = 1_000_000;
+    private static final long DIVISOR_KILO = 1_000_000;
     /** Divisor for megablob tier. */
-    private static final int DIVISOR_MEGA = 1_000_000_000;
+    private static final long DIVISOR_MEGA = 1_000_000_000;
     /** Number of significant digits in standard display format. */
     private static final int DISPLAY_SIG_DIGITS = 4;
     /** Number of significant digits in compact display format. */
@@ -37,15 +37,15 @@ public final class GooFormat {
     /** Space separator between value and suffix. */
     private static final String SEP_SPACE = " ";
 
-    /** Tiers in ascending order; last tier uses MAX_VALUE as a sentinel. */
+    /** Tiers in ascending order; the mega tier holds every long past the kilo tier. */
     private static final FormatTier[] TIERS = {
         new FormatTier(THRESHOLD_BLOB, DIVISOR_BLOB, SUFFIX_BLOB),
         new FormatTier(THRESHOLD_KILO, DIVISOR_KILO, SUFFIX_KILO),
-        new FormatTier(Integer.MAX_VALUE, DIVISOR_MEGA, SUFFIX_MEGA),
+        new FormatTier(Long.MAX_VALUE, DIVISOR_MEGA, SUFFIX_MEGA),
     };
 
     /** Volume tier: threshold, divisor, and suffix for a formatting bracket. */
-    private record FormatTier(int threshold, int divisor, String suffix) {}
+    private record FormatTier(long threshold, long divisor, String suffix) {}
 
     private GooFormat() {}
 
@@ -56,7 +56,7 @@ public final class GooFormat {
      * @param microblobs the volume in microblobs
      * @return the formatted display string
      */
-    public static String formatFluidDisplay(int microblobs) {
+    public static String formatFluidDisplay(long microblobs) {
         if (microblobs < THRESHOLD_MICROBLOB) { return formatMicroblobs(microblobs); }
         FormatTier tier = findTier(microblobs);
         return formatWithDecimal(microblobs, tier.divisor, tier.suffix);
@@ -67,7 +67,7 @@ public final class GooFormat {
      * @param microblobs the volume in microblobs
      * @return the matching tier
      */
-    private static FormatTier findTier(int microblobs) {
+    private static FormatTier findTier(long microblobs) {
         for (FormatTier tier : TIERS) {
             if (microblobs < tier.threshold) { return tier; }
         }
@@ -80,7 +80,7 @@ public final class GooFormat {
      * @param microblobs the volume in microblobs
      * @return the compact formatted string
      */
-    public static String formatFluidDisplayCompact(int microblobs) {
+    public static String formatFluidDisplayCompact(long microblobs) {
         if (microblobs < THRESHOLD_MICROBLOB) {
             return formatMicroblobs(microblobs);
         }
@@ -93,7 +93,7 @@ public final class GooFormat {
      * @param microblobs the sub-blob volume (0-999)
      * @return the decimal fraction string
      */
-    static String formatMicroblobs(int microblobs) {
+    static String formatMicroblobs(long microblobs) {
         return DOT + String.format(MICROBLOB_FORMAT, microblobs);
     }
 
@@ -105,9 +105,9 @@ public final class GooFormat {
      * @param suffix  the unit suffix (e.g. "K", "M")
      * @return the formatted string
      */
-    static String formatWithDecimal(int value, int divisor, String suffix) {
-        int whole = value / divisor;
-        int remainder = value % divisor;
+    static String formatWithDecimal(long value, long divisor, String suffix) {
+        long whole = value / divisor;
+        long remainder = value % divisor;
         int decimalDigits = DISPLAY_SIG_DIGITS - Long.toString(whole).length();
         String sep = suffix.isEmpty() ? SUFFIX_BLOB : SEP_SPACE;
         if (decimalDigits <= 0 || remainder == 0) { return whole + sep + suffix; }
@@ -122,7 +122,7 @@ public final class GooFormat {
      * @param digits    the number of decimal digits
      * @return the formatted fraction string
      */
-    static String formatFraction(int remainder, int divisor, int digits) {
+    static String formatFraction(long remainder, long divisor, int digits) {
         String raw = Long.toString(scaleRemainder(remainder, divisor, digits));
         return padAndTrimZeros(raw, digits);
     }
@@ -134,8 +134,8 @@ public final class GooFormat {
      * @param digits    the number of decimal digits
      * @return the scaled integer representing the decimal fraction
      */
-    private static int scaleRemainder(int remainder, int divisor, int digits) {
-        int scaled = remainder;
+    private static long scaleRemainder(long remainder, long divisor, int digits) {
+        long scaled = remainder;
         for (int i = 0; i < digits; i++) { scaled *= DECIMAL_BASE; }
         return scaled / divisor;
     }
@@ -161,7 +161,7 @@ public final class GooFormat {
      * @param microblobs the volume in microblobs
      * @return the compact formatted string
      */
-    private static String formatCompactWithSigDigits(int microblobs) {
+    private static String formatCompactWithSigDigits(long microblobs) {
         FormatTier tier = findTier(microblobs);
         return compactSigDigits(microblobs, tier.divisor, tier.suffix);
     }
@@ -174,9 +174,9 @@ public final class GooFormat {
      * @param suffix  the unit suffix
      * @return the compact formatted string
      */
-    private static String compactSigDigits(int value, int divisor, String suffix) {
-        int whole = value / divisor;
-        int remainder = value % divisor;
+    private static String compactSigDigits(long value, long divisor, String suffix) {
+        long whole = value / divisor;
+        long remainder = value % divisor;
         int wholeDigits = Long.toString(whole).length();
         int decimalDigits = COMPACT_SIG_DIGITS - wholeDigits;
